[fleet] Cleaning-up test
