+++ .idea/libraries/org_jetbrains_annotations_Latest.xml
@@ -2,7 +2,7 @@
   <library name="org.jetbrains:annotations:Latest" type="repository">
     <properties maven-id="org.jetbrains:annotations:LATEST" />
     <CLASSES>
-      <root url="jar://$MAVEN_REPOSITORY$/org/jetbrains/annotations/23.0.0/annotations-23.0.0.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jetbrains/annotations/24.1.0/annotations-24.1.0.jar!/" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
