+++ tests/org/intellij/grammar/BnfTestSuite.java
@@ -27,7 +27,7 @@ public class BnfTestSuite extends TestCase {
       testSuite.addTestSuite(JFlexGenerationTest.class);
       testSuite.addTestSuite(JFlexParserTest.class);
       testSuite.addTestSuite(BnfParserTest.class);
-      testSuite.addTestSuite(BnfGeneratorTestImpl.class);
+      testSuite.addTestSuite(BnfGeneratorTest.class);
       testSuite.addTestSuite(FleetBnfGeneratorTest.class);
       testSuite.addTestSuite(ExpressionParserTest.class);
       testSuite.addTestSuite(BnfLivePreviewParserTest.class);
